backend/requirements.txt
# ENSPD LearnAI - Backend Dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.1
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
tensorflow>=2.16.0
google-generativeai>=0.3.0
groq>=0.4.0
requests>=2.31.0
joblib>=1.3.2
